Moving visualization into Debug namespace to reduce bundle size in build
